requirements.txt
bjoern==3.2.1
Flask==2.0.2
python-dotenv==0.19.2
python-gitlab==3.4.0
python-sonarqube-api==1.2.9
redis==4.1.3
requests==2.27.1
rq==1.10.1
rq-exporter==1.9.3
